chore: refactor-login-form-type
