+++ src/components/forms/login-form/LoginForm.tsx
@@ -1,43 +1,13 @@
-import { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { LoginFormValues, loginSchema } from '../../../hooks/validation/login';
 import EmailInput from '../../ui/EmailInput';
 import PasswordInput from '../../ui/PasswordInput';
 import Button from '../../ui/Button';
 import Spinner from '../../ui/Spinner';
+import { Type } from '../../../features/auth/login/type';
 
-const LoginForm: React.FC = () => {
-    const {
-        register,
-        handleSubmit,
-        formState: { errors },
-    } = useForm<LoginFormValues>({
-        resolver: zodResolver(loginSchema),
-    });
 
-    const [loading, setLoading] = useState(false);
-    const [errorMsg, setErrorMsg] = useState('');
-
-    const onSubmit = async (data: LoginFormValues) => {
-        setLoading(true);
-        setErrorMsg('');
-
-        try {
-            console.log('Submitting login form with data:', data);
-            await new Promise((resolve) => setTimeout(resolve, 1000));
-        } catch (err) {
-            setErrorMsg('Login failed. Please try again.');
-        } finally {
-            setLoading(false);
-        }
-    };
-
+const LoginForm: React.FC<Type> = ({ register, errors, loading, errorMsg, onSubmit }) => {
     return (
-        <form
-            onSubmit={handleSubmit(onSubmit)}
-            className="flex flex-col gap-5 bg-white p-8 rounded-xl shadow-lg w-full"
-        >
+        <form onSubmit={onSubmit} className="flex flex-col gap-5 bg-white p-8 rounded-xl shadow-lg w-full">
             <EmailInput
                 label="Email Address"
                 {...register('email')}
