--- src/components/forms/login-form/LoginForm.types.ts
@@ -1,4 +1,0 @@
-export type LoginFormTypes = {
-    email: string;
-    password: string;
-};
