+++ src/hooks/validation/login.ts
@@ -1,6 +1,5 @@
 import { z } from 'zod';
 import { messages } from "../../lib/messages";
-import { LoginFormTypes } from '../../components/forms/login-form/LoginForm.types';
 
 export const loginSchema = z.object({
     email: z.string()
@@ -15,4 +14,4 @@ export const loginSchema = z.object({
         .max(255, messages.error.max('Password', 255))
 });
 
-export type LoginFormValues = LoginFormTypes & z.infer<typeof loginSchema>;
+export type LoginFormValues = z.infer<typeof loginSchema>;
